refactor(semantics): make Z3Theory package private
